+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FTC16072/TestWiring.java
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.teamcode.FTC16072;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
+import org.firstinspires.ftc.teamcode.FTC16072.mechanisms.Mechanism;
 import org.firstinspires.ftc.teamcode.FTC16072.tests.QQTest;
 
 import java.util.List;
