+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FTC16072/mechanisms/Mechanism.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.FTC16072;
+package org.firstinspires.ftc.teamcode.FTC16072.mechanisms;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
